requirements.txt
# Database
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0


# Data processing
pandas>=2.1.1
numpy>=1.26.0
